Fix asynx_read hanging on QNX.
The Linux man page for poll() says that receiving POLLHUP means that
the other end was closed. But it does not say if this is mandatory
behavior according to POSIX or not. On QNX it seems not to act like
this. Instead it returns POLLIN and then a zero read. A zero read
also means that the other end was closed (this is valid also on
Linux), so just treat those two cases the same.

This fixes a busy loop in the async_read function on QNX, since POLLIN
and a subsequent zero read are constantly spinning the loop after the
process has closed.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>